refactor: extracting non-main parts into lib.rs
extracting constants within run function to be generally available in lib file as well as extracting function definitions from within the function to make it easier to understand where one function ends and the other one begins.
changing some comments into documentation comments
